refactor: 공통 StoryBook 컴포넌트 레이아웃 변경
