+++ src/components/common/buttomsheet.tsx
@@ -12,8 +12,9 @@ export const Buttomsheet = () => {
   const handleFullClose = () => setIsFullHeightOpen(false);
 
   return (
-    <div className='p-4'>
-      <div className='flex gap-4 p-4'>
+    <div className='p-8'>
+      <h1 className='text-2xl font-bold mb-8'>바텀시트 예제</h1>
+      <div className='flex gap-4 pr-4'>
         {/* 반 높이 바텀시트 열기 버튼 */}
         <Button
           variant='primary'
