+++ src/components/common/buttons.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 
 const Buttons = () => {
   return (
-    <>
+    <div className='p-8'>
       <h1 className='text-2xl font-bold mb-4'>버튼 예제</h1>
 
       {/* 큰 버튼 */}
@@ -150,7 +150,7 @@ const Buttons = () => {
           </Button>
         </div>
       </div>
-    </>
+    </div>
   );
 };
