+++ src/components/common/example.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+const Example = () => {
+  return (
+    <div className='colors'>
+      <div className='p-8'>
+        <h1 className='text-2xl font-bold mb-8'>Example</h1>
+        <div className='grid grid-cols-3 gap-4'>Example</div>
+      </div>
+    </div>
+  );
+};
+
+export default Example;
