+++ src/components/groupList/RandomImage.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Card, CardContent } from '../ui/card';
+import { Card, CardContent } from '../common/card';
 import { Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious } from '@/components/ui/carousel';
 import { useGroupRandomImageQuery } from '@/hooks/queries/byUse/useGroupQueries';
 import Autoplay from 'embla-carousel-autoplay';
